requirements.txt
azure-common==1.1.14
azure-nspkg==2.0.0
azure-servicebus==0.21.1
certifi==2018.4.16
chardet==3.0.4
idna==2.7
python-qpid-proton==0.24.0
requests==2.19.1
urllib3==1.23